CORE-317: move common items to common folder
